requirements.txt
PyQt5
python-mpv
requests
Pillow
google-api-python-client
google-auth-oauthlib
